modified folder structure, improved create business function, update imports
