+++ app/__init__.py
@@ -2,4 +2,5 @@ from flask import Flask
 
 app = Flask(__name__, instance_relative_config=True)
 
-from app.auth.views import register, create_business
+from app.auth.views import register
+from app.businesses.views import create_business
