+++ run.py
@@ -1,6 +1,5 @@
 from app import app
 from instance.config import app_config
-from app.auth.views import register, create_business
 
 if __name__ =='__main__':
     app.config.from_object(app_config['development'])
